+++ src/fwk/toolkit/undo.cpp
@@ -46,7 +46,7 @@ void UndoTransaction::add(Command * cmd)
 
 void UndoTransaction::undo()
 {
-    DBG_OUT("undo transaction %d cmd", m_operations.size());
+    DBG_OUT("undo transaction %lu cmd", (unsigned long)m_operations.size());
 // I have no idea why this do not work
 //    std::for_each(m_operations.rbegin(), m_operations.rend(),
 //                  boost::bind(&Command::undo, _1));
@@ -59,7 +59,7 @@ void UndoTransaction::undo()
 
 void UndoTransaction::redo()
 {
-    DBG_OUT("redo transaction %d cmd", m_operations.size());
+    DBG_OUT("redo transaction %lu cmd", (unsigned long)m_operations.size());
 // I have no idea why this do not work
 //    std::for_each(m_operations.begin(), m_operations.end(),
 //                  boost::bind(&Command::redo, _1));
